Combat
Combat rounds now trigger every six seconds. All players on the server are checked for whether they are inCombat. If they are, they attempt a number of attacks on their target based on their baseAttackBonus. Each successful attack deals damage, on death a player is dramatically teleported back to room 1.